03-python-hw-consul/requirements.txt
flask
prometheus_client
python-consul
